+++ src/external_propagate.cpp
@@ -14,6 +14,8 @@ void Internal::add_observed_var (int ilit) {
   if (idx >= (int64_t) relevanttab.size ())
     relevanttab.resize (1 + (size_t) idx, 0);
   unsigned &ref = relevanttab[idx];
+  if (flags (idx).unused())
+    declare_variable (idx);
   if (ref < UINT_MAX) {
     ref++;
     LOG ("variable %d is observed %u times", idx, ref);
@@ -30,10 +32,11 @@ void Internal::add_observed_var (int ilit) {
     REQUIRE (!conflict,
              "can not observe assigned variable during conflict analysis");
     const int assignment_level = var (ilit).level;
-    backtrack (assignment_level - 1);
+    backtrack_without_updating_phases (assignment_level - 1);
   } else if (level && fixed (ilit)) {
-    backtrack (0);
+    backtrack_without_updating_phases (0);
   }
+  activating_all_new_imported_literals ();
 }
 
 /*----------------------------------------------------------------------------*/
@@ -49,7 +52,7 @@ void Internal::remove_observed_var (int ilit) {
         !conflict,
         "can not unobserve assigned variable during conflict analysis");
     const int assignment_level = var (ilit).level;
-    backtrack (assignment_level - 1);
+    backtrack_without_updating_phases (assignment_level - 1);
   }
 
   assert (fixed (ilit) || !val (ilit));
@@ -92,13 +95,13 @@ void Internal::set_changed_val () {
       continue;
     if (var (idx).reason != external_reason)
       continue;
-    if (!changed_val) {
-      changed_val = idx;
+    if (!earliest_changed_val) {
+      earliest_changed_val = idx;
       continue;
     }
-    assert (val (changed_val));
-    if (var (idx).level < var (changed_val).level) {
-      changed_val = idx;
+    assert (val (earliest_changed_val));
+    if (var (idx).level < var (earliest_changed_val).level) {
+      earliest_changed_val = idx;
     }
   }
 }
@@ -1268,15 +1271,16 @@ bool Internal::get_merged_literals (std::vector<int> &eq_class) {
 
   if (!lit_level) {
     // Collect all the variables that are merged and mapped to that ilit
-    size_t e2i_size = external->e2i.size ();
     int ivar = abs (ilit);
-    for (size_t i = 0; i < e2i_size; i++) {
-      int other = abs (external->e2i[i]);
+    for (auto id : external->e2i) {
+      int o_elit = id.second;
+      int o_ilit = id.first;
+      int other = abs (o_elit);
       if (other == ivar) {
-        if (external->e2i[i] == ilit)
-          eq_class.push_back (i);
+        if (o_elit == ilit)
+          eq_class.push_back (o_ilit);
         else
-          eq_class.push_back (-1 * i);
+          eq_class.push_back (-o_ilit);
       }
     }
 
@@ -1298,10 +1302,9 @@ void Internal::get_all_fixed_literals (std::vector<int> &fixed_lits) {
   if (!trail.size ())
     return;
 
-  int e2i_size = external->e2i.size ();
-  int ilit;
-  for (int eidx = 1; eidx < e2i_size; eidx++) {
-    ilit = external->e2i[eidx];
+  for (auto id : external->e2i) {
+    int ilit = id.second;
+    int eidx = id.first;
     if (ilit && !external->ervars[eidx]) {
       Flags &f = flags (ilit);
       if (f.status == Flags::FIXED) {