+++ src/factor.cpp
@@ -768,6 +768,7 @@ void Internal::schedule_factorization (Factoring &factoring) {
 
 void Internal::adjust_scores_and_phases_of_fresh_variables (
     Factoring &factoring) {
+  activating_all_new_imported_literals();
   if (!opts.factorunbump) {
     factoring.fresh.clear ();
     return;
@@ -792,13 +793,15 @@ void Internal::adjust_scores_and_phases_of_fresh_variables (
 #endif
 
   for (auto lit : factoring.fresh) {
-    LOG ("dequeuing %s", LOGLIT (lit));
-    queue.dequeue (links, lit);
+    LOG ("dequeuing %s, last being %s", LOGLIT (lit), LOGLIT (queue.last));
+    if (!flags (lit).fixed ())
+      queue.dequeue (links, lit);
   }
 
   for (auto lit : factoring.fresh) {
     LOG ("dequeuing %s", LOGLIT (lit));
-    queue.bury (links, lit);
+    if (!flags (lit).fixed ())
+      queue.bury (links, lit);
   }
 
   // fix the scores with negative numbers
@@ -813,7 +816,7 @@ void Internal::adjust_scores_and_phases_of_fresh_variables (
 
 #ifndef NDEBUG
   for (auto v : vars)
-    assert (val (v) || scores.contains (v));
+    assert (!flags (v).active () || val (v) || scores.contains (v));
   lit = queue.first;
   int next_lit = links[lit].next;
   while (next_lit) {