+++ src/propagate.cpp
@@ -104,6 +104,7 @@ inline void Internal::search_assign (int lit, Clause *reason) {
   if (level)
     require_mode (SEARCH);
 
+  assert (!flags (lit).unused());
   const int idx = vidx (lit);
   const bool from_external = reason == external_reason;
   assert (!val (idx));
@@ -229,6 +230,11 @@ bool Internal::propagate () {
     require_mode (SEARCH);
   assert (!unsat);
   LOG ("starting propagate");
+
+  // IPASIR-UP make everything harder.
+  if (!imports.empty())
+    activating_all_new_imported_literals();
+
   START (propagate);
 
   // Updating statistics counter in the propagation loops is costly so we