+++ src/util.hpp
@@ -3,6 +3,7 @@
 
 #include <cassert>
 #include <cstdint>
+#include <unordered_map>
 #include <vector>
 
 namespace CaDiCaL {
@@ -130,6 +131,13 @@ template <class T> static void reserve_at_least (vector<T> &v, size_t N) {
   v.reserve (new_size);
 }
 
+template <class K, class E>
+static K find_or_default (const std::unordered_map<K, E> &map, K key, E default_el) {
+  auto it = map.find (key);
+  if (it == map.end ())
+    return default_el;
+  return it->second;
+}
 // Clean-up class for bad_alloc error safety.
 
 template <typename T> struct DeferDeleteArray {